+++ src/lib/string.h
@@ -3,9 +3,9 @@
 
 #include <stdint.h>
 
-static uint32_t l_hash_string(const char* key, int length) {
+static uint32_t l_hash_string(const char* key, size_t length) {
     uint32_t hash = 2166136261u;
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
         hash ^= (uint8_t)key[i];
         hash *= 16777619;
     }
